requirements.txt
aniso8601==9.0.1
atomicwrites==1.4.0
attrs==21.2.0
Cerberus==1.3.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
clickclick==20.10.2
colorama==0.4.4
connexion==1.1.15
cryptography==3.4.8
flasgger==0.9.5
Flask==2.0.1
Flask-Cors==3.0.10
flask-restplus==0.13.0
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.8.1
inflection==0.5.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
mistune==0.8.4
packaging==21.0
pluggy==1.0.0
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
pytest-env==0.6.2
python-dateutil==2.6.0
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.23
SQLAlchemy-Utils==0.37.8
swagger-spec-validator==2.7.3
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.6
Werkzeug==2.0.1
wincertstore==0.2
zipp==3.5.0
pymysql
PyJWT